fix(structure): hooks directory restructuring
