+++ client/TODO.md
@@ -11,7 +11,7 @@
   - ~~Move from service classes to separate functions~~ (done: 19-01-26)
   - ~~Move from Axios to Ky~~ (done: 19-01-26)
   - Replace the spinner with one from the loading toast
-- Add titles to pages
+- ~~Add titles to pages~~ (done: 19-01-26)
 
 ## 03-01-26
