+++ client/src/components/features/articles/ArticleCreateForm.tsx
@@ -11,7 +11,7 @@ import {
 } from '@/components/ui/dialog';
 import { Button } from '@/components/ui/button';
 import { Input } from '@/components/ui/input';
-import { useCreateArticle } from '@/hooks/articles/useCreateArticle';
+import { useCreateArticle } from '@/hooks/features/articles/useCreateArticle';
 import type { ArticleCreate } from '^/types/articles';
 import { FormField } from '@/components/common/form/FormField';
 import { Spinner } from '@/components/common/spinner/Spinner';
