+++ client/src/components/features/auth/forms/ForgotPasswordForm.tsx
@@ -2,7 +2,7 @@ import { useTranslation } from 'react-i18next';
 import { useCloseDialog } from '@/hooks/ui/useCloseDialog';
 import { useState } from 'react';
 import { Button } from '@/components/ui/button';
-import { usePasswordResetRequest } from '@/hooks/auth/password-reset/use-request-password-reset';
+import { usePasswordResetRequest } from '@/hooks/features/auth/password-reset/use-request-password-reset';
 import { useForm, type SubmitHandler } from 'react-hook-form';
 import { yupResolver } from '@hookform/resolvers/yup';
 import { authPwResetRequestSchema } from '^/schemas/auth/auth-pw-reset-request.schema';
