+++ client/src/hooks/features/auth/useLogout.ts
