+++ client/src/providers/AuthProvider.tsx
@@ -1,5 +1,5 @@
 import { useEffect, type PropsWithChildren } from 'react';
-import { useRefresh } from '@/hooks/auth/useRefresh';
+import { useRefresh } from '@/hooks/features/auth/useRefresh';
 import { ACCESS_TOKEN } from '^/constants/auth.constants';
 
 export const AuthProvider = ({ children }: PropsWithChildren) => {
